RAG/multimodal-rag/requirements.txt
chromadb
Pillow
matplotlib
datasets
python-dotenv
langchain
langchain-core
langchain-community
langchain-ollama
streamlit
ollama
